requirements.txt
pymongo
requests
ipython
python-dotenv
pandas